refactor: split price filter
